fix(TASK-1004): remove fakeData because not a test and use its content inside SubscriptionDialog.test.ts
